Move shared components out of app/ directory
