+++ web/src/components/search.tsx
@@ -3,7 +3,7 @@ import React, { useState, useEffect } from "react";
 import ReactSelect from "react-select";
 import { useRouter } from "next/navigation";
 import { MagnifierRight } from "./icons";
-import { SearchResult } from "../api/search/route";
+import { SearchResult } from "../app/api/search/route";
 
 export const SearchInput = () => {
   const [results, setCities] = useState<SearchResult[]>([]);
